% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{check_if_on_cran}
\alias{check_if_on_cran}
\title{Check if package is available on CRAN.}
\usage{
check_if_on_cran(remote_ref, version = NULL)
}
\description{
Check if package is available on CRAN.
}
\examples{
check_if_on_cran(list(package = "magrittr"))
check_if_on_cran(list(package = "magrittr"), list(op = ">=", op_ver = "0.5.0"))
check_if_on_cran(list(package = "magrittr"), list(op = ">=", op_ver = "9999.9.99"))
check_if_on_cran(list(package = "magrittr"), list(op = "<", op_ver = "0.0.0"))
}
\keyword{internal}
