% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_ip}
\alias{check_ip}
\title{Executes \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}} on a local package using \code{libpath} from the installation plan.}
\usage{
check_ip(ip, path, build_args = character(), check_args = character(), ...)
}
\arguments{
\item{ip}{(\code{pkg_installation_plan}) object to execute}

\item{path}{(\code{string}) path to the package sources}

\item{build_args}{(\code{string}) value passed as \code{build_args} argument into \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}}

\item{check_args}{(\code{string}) value passed as \code{args} argument into \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}}

\item{...}{other arguments passed to \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}}
}
\value{
An S3 object (list) with fields \code{errors},
\code{warnings} and \code{notes}. These are all character
vectors containing the output for the failed check.
}
\description{
Executes \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}} on a local package using \code{libpath} from the installation plan.
}
\seealso{
\code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}} for other configuration possibilities
}
