% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_installation_proposal.R
\name{cli_pb_init}
\alias{cli_pb_init}
\title{Create \code{cli} progress bar for resolving versions.}
\usage{
cli_pb_init(type, total, ...)
}
\description{
Create \code{cli} progress bar for resolving versions.
}
\keyword{internal}
