% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{deparse_dep_str}
\alias{deparse_dep_str}
\title{Deparse a dependency string}
\usage{
deparse_dep_str(x)
}
\arguments{
\item{x}{(\code{character}) Dependency string in form of \verb{<package name> (<operator> <version>)}. See examples.}
}
\value{
\code{list} with \code{package}, \code{op}, \code{op_ver} and \code{ver_str} fields.
}
\description{
Deparse a dependency string
}
\examples{
verdepcheck:::deparse_dep_str("foo")
verdepcheck:::deparse_dep_str("foo (>= 1.2.3)")
}
\keyword{internal}
