% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{deparse_ver_str}
\alias{deparse_ver_str}
\title{Deparse a version string}
\usage{
deparse_ver_str(x)
}
\arguments{
\item{x}{(\code{character}) Version string in form of \verb{<operator> <version>}. See examples.}
}
\value{
\code{list} with \code{op} and \code{op_ver} fields.
}
\description{
Deparse a version string
}
\examples{
verdepcheck:::deparse_ver_str(">= 1.2.3")
}
\keyword{internal}
