% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{max_deps_check}
\alias{max_deps_check}
\alias{release_deps_check}
\alias{min_direct_deps_check}
\alias{min_cohort_deps_check}
\alias{min_cohorts_deps_check}
\title{Execute \verb{R CMD CHECK} on a local package with all dependencies pre-installed using various strategies.}
\usage{
max_deps_check(
  path,
  config = list(),
  build_args = character(),
  check_args = character(),
  ...
)

release_deps_check(
  path,
  config = list(),
  build_args = character(),
  check_args = character(),
  ...
)

min_direct_deps_check(
  path,
  config = list(),
  build_args = character(),
  check_args = character(),
  ...
)

min_cohort_deps_check(
  path,
  config = list(),
  build_args = character(),
  check_args = character(),
  ...
)

min_cohorts_deps_check(
  path,
  config = list(),
  build_args = character(),
  check_args = character(),
  ...
)
}
\arguments{
\item{path}{(\code{string}) path to the package sources}

\item{config}{(\code{list}) configuration options. See \code{\link[pkgdepends:pkg_config]{pkgdepends::pkg_config}} for details.
\code{"dependencies"} and \code{"library"} elements are overwritten by package level defaults.}

\item{build_args}{(\code{string}) value passed as \code{build_args} argument into \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}}

\item{check_args}{(\code{string}) value passed as \code{args} argument into \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}}

\item{...}{other arguments passed to \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}}
}
\value{
a named \code{list} with two elements:
\itemize{
\item \code{"ip"} - installation plan object
\item \code{"check"} - returned value from \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}
}
}
\description{
Execute \verb{R CMD CHECK} on a local package with all dependencies pre-installed using various strategies.
}
\section{strategies}{

Currently implemented strategies:
\itemize{
\item \code{max} - use the greatest version of dependent packages. Please note that using development version is not
guaranteed to be stable.
See \link{get_ref_max} for details.
\item \code{release} - use the released version of dependent packages. It will try use CRAN if possible else if
GitHub release is available then use it else fail.
See \link{get_ref_release} for details.
\item \code{min_direct} - use the lowest version of direct dependent packages that satisfy version condition.
If no version is specified then the minimal available version is assumed.
Indirect dependencies are installed as usual, i.e. using the greatest available.
See \link{get_ref_min} for details.
\item \code{min_cohort} - find maximum date of directly dependent packages release dates and use that as PPM snapshot date
for dependency resolve.
This will limit the versions of indirect dependencies by date of parent package release date.
\item \code{min_cohorts} - for each direct dependency: find its release date and use it as PPM snapshot for that dependency
resolve. Next, combine all the individual resolve outputs and resolve it again.
This will limit the versions of indirect dependencies by date of parent package release date.
}

Please note that only \code{min_cohort} and \code{min_cohorts} strategies are "stable" (unless "@*release" dynamic
pointer is used in any of the dependencies - including also indirect ones).
For all other results might differ due to changes in dependencies including a release of a new version.
The most straightforward example is \code{max} strategy in which the environment will be different after any push of
any of the dependencies.
}

\examples{
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- max_deps_check(".")
x$ip
x$check
\dontshow{\}) # examplesIf}
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- release_deps_check(".")
x$ip
x$check
\dontshow{\}) # examplesIf}
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- min_direct_deps_check(".")
x$ip
x$check
\dontshow{\}) # examplesIf}
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- min_cohort_deps_check(".")
x$ip
x$check
\dontshow{\}) # examplesIf}
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- min_cohorts_deps_check(".")
x$ip
x$check
\dontshow{\}) # examplesIf}
}
\seealso{
\link{deps_installation_proposal}
}
