% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{desc_add_extra_deps}
\alias{desc_add_extra_deps}
\title{Adds extra dependencies to the \code{desc} object.}
\usage{
desc_add_extra_deps(d, x)
}
\arguments{
\item{d}{(\code{desc}) DESCRIPTION object from \link[desc:desc]{desc::desc}}

\item{x}{(\code{character(1)}) Extra dependencies specified similarly to the \code{DESCRIPTION} file, e.g.
\code{"<package name> (<operator> <version>)"} where both \code{operator} and \code{version} are optional.
Multiple entries are possible separated by \code{";"}.}
}
\value{
\code{desc} object with added dependencies.
}
\description{
This will add dependencies to the \code{Imports} field.
}
\examples{
d <- desc::desc(cmd = "!new")
d <- verdepcheck:::desc_add_extra_deps(d, "foo (>= 1.2.3); bar (== 2.3.4)")
d$get_deps()
}
\keyword{internal}
