% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{desc_cond_set_refs}
\alias{desc_cond_set_refs}
\title{Set \code{"Config/Needs/verdepcheck"} section into the \code{desc} object if not empty else clear this section.}
\usage{
desc_cond_set_refs(d, refs)
}
\description{
Set \code{"Config/Needs/verdepcheck"} section into the \code{desc} object if not empty else clear this section.
}
\keyword{internal}
