% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_installation_proposal.R
\name{desc_cond_set_remotes}
\alias{desc_cond_set_remotes}
\title{Set \code{"Remotes"} section into the \code{desc} object if not empty else clear \code{"Remotes"} section.}
\usage{
desc_cond_set_remotes(desc, remotes)
}
\description{
Set \code{"Remotes"} section into the \code{desc} object if not empty else clear \code{"Remotes"} section.
}
\keyword{internal}
