% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{desc_remotes_cleanup}
\alias{desc_remotes_cleanup}
\title{Replace Remotes in the \code{desc} that have been resolved to a GitHub tag or are
in CRAN}
\usage{
desc_remotes_cleanup(d)
}
\arguments{
\item{d}{(\code{desc}) DESCRIPTION object}
}
\description{
Replaces any existing Remotes entry with the resolved GitHub tag from
\code{Config/Needs/verdepcheck}.
}
\examples{
# Example that should replace rtables & formatters on Remotes
#  but not pkgdepends

d <- desc::desc(
  file = verdepcheck:::local_description(
    list(
      rtables = "Import", formatters = "Import", pkgdepends = "Import",
      dplyr = "Import"
    ),
    remotes = c(
      "insightsengineering/rtables@*release",
      "insightsengineering/formatters@*release",
      "r-lib/pkgdepends@*release"
    ),
    need_verdepcheck = c(
      "dplyr",
      "rtables=insightsengineering/rtables@0.6.2",
      "formatters=insightsengineering/formatters@0.5.1"
    )
  )
)
verdepcheck:::desc_remotes_cleanup(d)
}
\keyword{internal}
