% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{desc_to_ip}
\alias{desc_to_ip}
\title{Create \code{installation_plan} object from \code{desc} object}
\usage{
desc_to_ip(d, config)
}
\description{
Create \code{installation_plan} object from \code{desc} object
}
\keyword{internal}
