% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{download_ip}
\alias{download_ip}
\title{Solve installation plan ignoring entries with "@*release" remote refs for detected conflicts.}
\usage{
download_ip(ip)
}
\arguments{
\item{ip}{(\code{pkg_installation_plan}) object to execute}
}
\value{
\code{pkg_installation_plan} object invisibly
}
\description{
Solve installation plan ignoring entries with "@*release" remote refs for detected conflicts.
}
