% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enforce_rcpp}
\alias{enforce_rcpp}
\title{Enforce a minimum version of Rcpp (>=1.0.0) for R version above 4.0.0
A change in base R on 4.0.0 makes Rcpp incompatible in previous versions}
\usage{
enforce_rcpp(pkg_resolution)
}
\description{
Enforce a minimum version of Rcpp (>=1.0.0) for R version above 4.0.0
A change in base R on 4.0.0 makes Rcpp incompatible in previous versions
}
\keyword{internal}
