% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{execute_ip}
\alias{execute_ip}
\title{Executes installation plan and \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}} in "try mode" to always return.}
\usage{
execute_ip(ip, path, build_args, check_args, ...)
}
\arguments{
\item{ip}{(\code{pkg_installation_plan}) object to execute}

\item{path}{(\code{string}) path to the package sources}

\item{build_args}{(\code{string}) value passed as \code{build_args} argument into \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}}

\item{check_args}{(\code{string}) value passed as \code{args} argument into \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}}

\item{...}{other arguments passed to \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}}
}
\value{
a named \code{list} with two elements:
\itemize{
\item \code{"ip"} - installation plan object
\item \code{"check"} - returned value from \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}
}
}
\description{
Executes installation plan and \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}} in "try mode" to always return.
}
