% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{filter_valid_version}
\alias{filter_valid_version}
\title{Filter for package versions that comply with an operator and version}
\usage{
filter_valid_version(x, op, op_ver)
}
\arguments{
\item{x}{\code{vector} of valid package versions.}

\item{op}{\code{character(1)} relational operator (\code{>=}, \code{==}, ...)}

\item{op_ver}{\code{character(1)} or \code{package_version(1)} with version to compare
with using a relational operator.}
}
\description{
Filter for package versions that comply with an operator and version
}
\examples{
versions <- paste(1:10, 0, sep = ".")
verdepcheck:::filter_valid_version(versions, ">=", "3.1")
}
\keyword{internal}
