% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_minver_remote_ref.R
\name{find_minver_remote_ref}
\alias{find_minver_remote_ref}
\alias{find_minver_remote_ref.remote_ref}
\alias{find_minver_remote_ref.remote_ref_cran}
\alias{find_minver_remote_ref.remote_ref_standard}
\alias{find_minver_remote_ref.remote_ref_github}
\title{Identify minimal version.}
\usage{
find_minver_remote_ref(remote_ref, op = "", op_ver = "")

\method{find_minver_remote_ref}{remote_ref}(remote_ref, op = "", op_ver = "")

\method{find_minver_remote_ref}{remote_ref_cran}(remote_ref, op = "", op_ver = "")

\method{find_minver_remote_ref}{remote_ref_standard}(remote_ref, op = "", op_ver = "")

\method{find_minver_remote_ref}{remote_ref_github}(remote_ref, op = "", op_ver = "")
}
\arguments{
\item{remote_ref}{(\code{remote_ref}) object created with \code{\link[pkgdepends:parse_pkg_refs]{pkgdepends::parse_pkg_ref()}}}

\item{op}{(\code{character(1)}) optional, version condition comparison operator (e.g. \code{">"}, \code{">="})}

\item{op_ver}{(\code{character(1)}) optional, version number against which \code{op} argument is applied}
}
\value{
(\code{remote_ref}) object with reference to the package of minimal version.
}
\description{
Identify minimal version.

Identify minimal version for CRAN-type of remote. This would use \code{\link[pkgcache:cran_archive_list]{pkgcache::cran_archive_list()}}
to obtain historical data.

Identify minimal version of GitHub type of remote.
}
\note{
only \code{cran}, \code{standard} (treated as \code{cran}) and \code{github} remote type are supported.
For the rest, it would return \code{remote_ref} argument.
}
\examples{
find_minver_remote_ref(pkgdepends::parse_pkg_ref("bioc::MultiAssayExperiment"))
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
find_minver_remote_ref(pkgdepends::parse_pkg_ref("cran::dplyr"))
\dontshow{\}) # examplesIf}
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
find_minver_remote_ref(pkgdepends::parse_pkg_ref("dplyr"))
\dontshow{\}) # examplesIf}
\dontshow{if (gh::gh_token() != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
find_minver_remote_ref(pkgdepends::parse_pkg_ref("cran/dplyr"))
\dontshow{\}) # examplesIf}
}
\seealso{
\code{gh::gh()} for other configuration possibilities
}
