% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_avail_date}
\alias{get_avail_date}
\alias{get_avail_date.remote_ref_cran}
\alias{get_avail_date.remote_ref_standard}
\alias{get_avail_date.remote_ref}
\title{Get available date for the package.}
\usage{
get_avail_date(remote_ref, start = get_release_date(remote_ref))

\method{get_avail_date}{remote_ref_cran}(remote_ref, start = get_release_date(remote_ref))

\method{get_avail_date}{remote_ref_standard}(remote_ref, start = get_release_date(remote_ref))

\method{get_avail_date}{remote_ref}(remote_ref, start = get_release_date(remote_ref))
}
\arguments{
\item{remote_ref}{(\code{remote_ref}) object created with \code{\link[pkgdepends:parse_pkg_refs]{pkgdepends::parse_pkg_ref()}}}

\item{start}{(\code{Date}) optional, the date when the package was released}
}
\value{
Date. This can be safely used as a snapshot date for the PPM.
}
\description{
Oftentimes, the release date of the package does not correspond to the date when the package is
available in the PPM. Usually it takes one day for the PPM to sync with the CRAN.
This function will return the date when the package is available in the PPM.
}
\examples{
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_avail_date(pkgdepends::parse_pkg_ref("cran::dplyr"))
get_avail_date(pkgdepends::parse_pkg_ref("cran::dplyr@1.1.0"))
\dontshow{\}) # examplesIf}
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_avail_date(pkgdepends::parse_pkg_ref("dplyr"))
get_avail_date(pkgdepends::parse_pkg_ref("dplyr@1.1.0"))
\dontshow{\}) # examplesIf}
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "" && gh::gh_token() != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_avail_date(pkgdepends::parse_pkg_ref("bioc::MultiAssayExperiment"))
get_avail_date(pkgdepends::parse_pkg_ref("tidyverse/dplyr@v1.1.0"))
\dontshow{\}) # examplesIf}
}
