% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_bioc_package_release_date}
\alias{get_bioc_package_release_date}
\title{Get the pseudo-published date of Bioconductor packages}
\usage{
get_bioc_package_release_date(package)
}
\description{
The date from the last modification to the sources is assumed to be the
published date as it is not available in \code{pkgcache} results.
}
\details{
The official metadata of the Bioconductor release does not work for the
purpose of this package (\verb{https://bioconductor.org/config.yaml}). As there
might be a edge case where a Bioconductor package was changed and depends
on a newer version of a package. (Actual edge case with \code{SummarizedExperiment}
and \code{matrixStats})
}
\keyword{internal}
