% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{get_desc_field_pkgs}
\alias{get_desc_field_pkgs}
\title{Get the packages from the custom config field}
\usage{
get_desc_field_pkgs(d)
}
\arguments{
\item{d}{(\code{desc}) DESCRIPTION object from \link[desc:description]{desc::description}}
}
\value{
character string
}
\description{
Get the packages from the custom config field
}
\keyword{internal}
