% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_desc_from_gh}
\alias{get_desc_from_gh}
\title{Get DESCRIPTION from GitHub Repository}
\usage{
get_desc_from_gh(org, repo, ref = "")
}
\description{
Get DESCRIPTION from GitHub Repository
}
\examples{
\dontshow{if (gh::gh_token() != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
verdepcheck:::get_desc_from_gh("tidyverse", "dplyr")
verdepcheck:::get_desc_from_gh("tidyverse", "dplyr", "v1.1.0")
\dontshow{\}) # examplesIf}
}
\keyword{internal}
