% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min_ver.R
\name{get_min_ver}
\alias{get_min_ver}
\alias{get_min_ver.remote_ref}
\alias{get_min_ver.remote_ref_cran}
\alias{get_min_ver.remote_ref_standard}
\alias{get_min_ver.remote_ref_github}
\title{Get minimal version.}
\usage{
get_min_ver(remote_ref, op = "", op_ver = "")

\method{get_min_ver}{remote_ref}(remote_ref, op = "", op_ver = "")

\method{get_min_ver}{remote_ref_cran}(remote_ref, op = "", op_ver = "")

\method{get_min_ver}{remote_ref_standard}(remote_ref, op = "", op_ver = "")

\method{get_min_ver}{remote_ref_github}(remote_ref, op = "", op_ver = "")
}
\arguments{
\item{remote_ref}{(\code{remote_ref}) object created with \code{\link[pkgdepends:parse_pkg_refs]{pkgdepends::parse_pkg_ref()}}}

\item{op}{(\code{character(1)}) optional, version condition comparison operator (e.g. \code{">"}, \code{">="})}

\item{op_ver}{(\code{character(1)}) optional, version number against which \code{op} argument is applied}
}
\value{
(\code{remote_ref}) object with reference to the package of minimal version.
}
\description{
Get minimal version.

\itemize{
\item for CRAN-type of remote - this would use \code{\link[pkgcache:cran_archive_list]{pkgcache::cran_archive_list()}}
to obtain historical data.
}

\itemize{
\item for GitHub type of remote - this would use \code{\link[gh:gh_gql]{gh::gh_gql()}} to get list of all tags
and then \code{\link[gh:gh]{gh::gh()}} to download \code{DESCRIPTION} file and then read package version.
}
}
\examples{
get_min_ver(pkgdepends::parse_pkg_ref("bioc::MultiAssayExperiment"))
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_min_ver(pkgdepends::parse_pkg_ref("cran::dplyr"))
\dontshow{\}) # examplesIf}
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_min_ver(pkgdepends::parse_pkg_ref("dplyr"))
\dontshow{\}) # examplesIf}
\dontshow{if (gh::gh_token() != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_min_ver(pkgdepends::parse_pkg_ref("cran/dplyr"))
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=get_min_ver_incl_cran]{get_min_ver_incl_cran()}}
}
