% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min_ver.R
\name{get_min_ver_incl_cran}
\alias{get_min_ver_incl_cran}
\alias{get_min_ver_incl_cran.remote_ref}
\alias{get_min_ver_incl_cran.remote_ref_github}
\title{Get minimal version including also check in CRAN repository.}
\usage{
get_min_ver_incl_cran(remote_ref, op = "", op_ver = "")

\method{get_min_ver_incl_cran}{remote_ref}(remote_ref, op = "", op_ver = "")

\method{get_min_ver_incl_cran}{remote_ref_github}(remote_ref, op = "", op_ver = "")
}
\arguments{
\item{remote_ref}{(\code{remote_ref}) object created with \code{\link[pkgdepends:parse_pkg_refs]{pkgdepends::parse_pkg_ref()}}}

\item{op}{(\code{character(1)}) optional, version condition comparison operator (e.g. \code{">"}, \code{">="})}

\item{op_ver}{(\code{character(1)}) optional, version number against which \code{op} argument is applied}
}
\value{
(\code{remote_ref}) object with reference to the package of minimal version.
}
\description{
Get minimal version including also check in CRAN repository.
}
\examples{
get_min_ver_incl_cran(pkgdepends::parse_pkg_ref("bioc::MultiAssayExperiment"))
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_min_ver_incl_cran(pkgdepends::parse_pkg_ref("cran::dplyr"))
\dontshow{\}) # examplesIf}
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_min_ver_incl_cran(pkgdepends::parse_pkg_ref("cran/dplyr"))
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=get_min_ver_incl_cran]{get_min_ver_incl_cran()}}
}
