% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_ref_internal}
\alias{get_ref_internal}
\title{Get reference to the release or maximal version of the package.}
\usage{
get_ref_internal(remote_ref, add_gh_releases)
}
\arguments{
\item{remote_ref}{(\code{remote_ref}) object created with \code{\link[pkgdepends:parse_pkg_refs]{pkgdepends::parse_pkg_ref()}}}

\item{add_gh_releases}{(\code{logical(1)}) if \code{TRUE} then also check GitHub releases}
}
\value{
(\code{remote_ref}) object with the package reference
}
\description{
Get reference to the release or maximal version of the package.
}
\keyword{internal}
