% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_ref_release}
\alias{get_ref_release}
\title{Get reference to the release version of the package.}
\usage{
get_ref_release(remote_ref)
}
\arguments{
\item{remote_ref}{(\code{remote_ref}) object created with \code{\link[pkgdepends:parse_pkg_refs]{pkgdepends::parse_pkg_ref()}}}
}
\value{
(\code{remote_ref}) object with the package reference
}
\description{
Get reference to the release version of the package.
}
\examples{
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_ref_release(pkgdepends::parse_pkg_ref("dplyr"))
get_ref_release(pkgdepends::parse_pkg_ref("tidyverse/dplyr"))
\dontshow{\}) # examplesIf}
}
