% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_installation_proposal.R
\name{get_refs_from_desc}
\alias{get_refs_from_desc}
\title{Read \code{"Config/Needs/verdepcheck"} section and return vector of references.}
\usage{
get_refs_from_desc(d)
}
\description{
Read \code{"Config/Needs/verdepcheck"} section and return vector of references.
}
\keyword{internal}
