% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_release_data}
\alias{get_release_data}
\title{Get data for CRAN/Bioconductor package releases}
\usage{
get_release_data(package)
}
\description{
Get data for CRAN/Bioconductor package releases
}
\examples{
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
verdepcheck:::get_release_data("dplyr")
verdepcheck:::get_release_data("MultiAssayExperiment")
\dontshow{\}) # examplesIf}
}
\keyword{internal}
