% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_release_date}
\alias{get_release_date}
\alias{get_release_date.remote_ref_github}
\alias{get_release_date.remote_ref_cran}
\title{Get release date.}
\usage{
get_release_date(remote_ref)

\method{get_release_date}{remote_ref_github}(remote_ref)

\method{get_release_date}{remote_ref_cran}(remote_ref)
}
\arguments{
\item{remote_ref}{(\code{remote_ref}) object created with \code{\link[pkgdepends:parse_pkg_refs]{pkgdepends::parse_pkg_ref()}}}
}
\value{
Date
}
\description{
Get release date.

Get release date from GitHub references
}
\examples{
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "" && gh::gh_token() != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_release_date(pkgdepends::parse_pkg_ref("tidyverse/dplyr@v1.1.0"))
\dontshow{\}) # examplesIf}
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
remote_ref <- pkgdepends::parse_pkg_ref("dplyr@1.1.0")
get_release_date.remote_ref_cran(remote_ref)
\dontshow{\}) # examplesIf}
}
