% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_release_date.remote_ref_cran}
\alias{get_release_date.remote_ref_cran}
\title{Get release date from GitHub references}
\usage{
\method{get_release_date}{remote_ref_cran}(remote_ref)
}
\arguments{
\item{remote_ref}{(\code{remote_ref}) object created with \code{\link[pkgdepends:parse_pkg_refs]{pkgdepends::parse_pkg_ref()}}}
}
\description{
Get release date from GitHub references
}
\examples{
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
remote_ref <- pkgdepends::parse_pkg_ref("dplyr@1.1.0")
get_release_date.remote_ref_cran(remote_ref)
\dontshow{\}) # examplesIf}
}
