% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_release_date.remote_ref_github}
\alias{get_release_date.remote_ref_github}
\title{Get release date from GitHub references}
\usage{
\method{get_release_date}{remote_ref_github}(remote_ref)
}
\arguments{
\item{remote_ref}{(\code{remote_ref}) object created with \code{\link[pkgdepends:parse_pkg_refs]{pkgdepends::parse_pkg_ref()}}}
}
\description{
Get release date from GitHub references
}
\examples{
\dontshow{if (gh::gh_token() != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
remote_ref <- pkgdepends::parse_pkg_ref("tidyverse/dplyr@v1.1.0")
get_release_date(remote_ref)
\dontshow{\}) # examplesIf}
}
