% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_version}
\alias{get_version}
\alias{get_version.remote_ref}
\title{Get package version.}
\usage{
get_version(remote_ref)

\method{get_version}{remote_ref}(remote_ref)
}
\arguments{
\item{remote_ref}{(\code{remote_ref}) object created with \code{\link[pkgdepends:parse_pkg_refs]{pkgdepends::parse_pkg_ref()}}}
}
\value{
Package version created with \code{package_version}.
}
\description{
Get package version.
}
\examples{
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_version(pkgdepends::parse_pkg_ref("dplyr"))
get_version(pkgdepends::parse_pkg_ref("tidyverse/dplyr"))
get_version(pkgdepends::parse_pkg_ref("tidyverse/dplyr@v1.1.0"))
get_version(pkgdepends::parse_pkg_ref("bioc::MultiAssayExperiment"))
\dontshow{\}) # examplesIf}
}
