% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{install_ip}
\alias{install_ip}
\title{Executes installation plan.}
\usage{
install_ip(ip)
}
\arguments{
\item{ip}{(\code{pkg_installation_plan}) object to execute}
}
\value{
\code{pkg_installation_plan} object invisibly
}
\description{
This function would executes the following:
\itemize{
\item solves package dependencies
\item downloads all package dependencies
\item installs system requirements
\item installs all package dependencies
}
}
\keyword{internal}
