% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_individual_isolated.R
\name{install_pkg_versions_tmp}
\alias{install_pkg_versions_tmp}
\title{Test install all the versions of the package locally}
\usage{
install_pkg_versions_tmp(
  package,
  op = "",
  op_ver = "",
  max_count = NA,
  until_success = TRUE
)
}
\arguments{
\item{package}{(\code{character}) name of R package.}

\item{op}{(\code{character}) relational operator.}

\item{op_ver}{(\code{character}) version of package.}

\item{max_count}{(\code{integer}) maximum number of versions to test, \code{NA} for no
limit.}

\item{until_success}{(\code{logical}) flag that indicates whether a
successful installation of a package should stop the search.}
}
\description{
Support function to assess if every version specified by \code{op} and \code{op_ver}
of the package can be installed. This is not always the case as there might
be some API changes that break specific versions.
}
\details{
The \code{Rcpp} package before 0.12.16 cannot be compiled by R versions above 4.0

There is no output for this function other than the messages in the console.
}
\examples{
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
install_pkg_versions_tmp("checkmate", ">=", "2.1.0")
install_pkg_versions_tmp("Rcpp", ">=", "0.12.15", 3, FALSE)
install_pkg_versions_tmp("Rcpp", ">=", "0.12.15", NA, TRUE)

# Make sure the latest version on CRAN is working
dplyr_latest <- pkgcache::meta_cache_list("dplyr")$version[1]
install_pkg_versions_tmp("dplyr", ">=", dplyr_latest)
\dontshow{\}) # examplesIf}
}
