% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_description}
\alias{local_description}
\title{Temporarily create a valid DESCRIPTION file to a location that will be deleted}
\usage{
local_description(
  pkg_list = c(pkgdepends = "Import"),
  remotes = c(),
  need_verdepcheck = c(),
  .local_envir = parent.frame()
)
}
\arguments{
\item{pkg_list}{(\code{vector}) named character vector or list with
paired name and type of dependency. It supports versions by using quotes on
the key}

\item{remotes}{(\code{vector}) string vector that contains remotes to add to
the DESCRIPTION file}

\item{need_verdepcheck}{(\code{vector}) string vector that contains
Config/Need/verdepcheck elements to add to the DESCRIPTION file}

\item{.local_envir}{(\code{envirnoment}) The environment to use for scoping.}
}
\description{
The file is deleted after the parent environment where this function was called
has exited, when the R session ends or on deman via \code{\link[withr:defer]{withr::deferred_run()}}
}
\examples{
verdepcheck:::local_description(
  list(rtables = "Import"),
  remotes = "insightsengineering/rtables",
  need_verdepcheck = "rtables=insightsengineering/rtables@0.6.2"
)
}
\keyword{internal}
