% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_key_character}
\alias{map_key_character}
\title{Parse through vector of \code{remote_ref} and retrieve one of the keys of each
element}
\usage{
map_key_character(x, key)
}
\arguments{
\item{x}{(\code{list}) list of lists where each internal list contain the same key}

\item{field}{(\code{character(1)}) key of field to ret}
}
\description{
Support function to reduce repetive code
}
\examples{
verdepcheck:::map_key_character(
  list(list(a = "1", b = "2"), list(a = "3", b = "4"), list(a = "5", b = "6")),
  "a"
)
}
\keyword{internal}
