% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{resolve_ignoring_release_remote}
\alias{resolve_ignoring_release_remote}
\title{If solution errors finishes with "dependency conflict" error then
re-try again ignoring "@*release" remote refs for detected conflicts.}
\usage{
resolve_ignoring_release_remote(ip)
}
\arguments{
\item{ip}{(\code{pkg_installation_plan}) object to execute}
}
\value{
\code{pkg_installation_plan} object invisibly
}
\description{
If solution errors finishes with "dependency conflict" error then
re-try again ignoring "@*release" remote refs for detected conflicts.
}
\keyword{internal}
