% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_ppm_snapshot}
\alias{resolve_ppm_snapshot}
\title{Resolve the dependencies of a package based on its release date + 1.}
\usage{
resolve_ppm_snapshot(pkg_ref_str, operator, pkg_version)
}
\description{
Resolve the dependencies of a package based on its release date + 1.
}
\keyword{internal}
