% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{solve_ignore_remotes_release}
\alias{solve_ignore_remotes_release}
\title{Solve installation plan ignoring entries with "@*release" remote refs for detected conflicts.}
\usage{
solve_ignore_remotes_release(ip)
}
\arguments{
\item{ip}{(\code{pkg_installation_plan}) object to execute}
}
\value{
\code{pkg_installation_plan} object invisibly
}
\description{
Solve installation plan ignoring entries with "@*release" remote refs for detected conflicts.
}
\keyword{internal}
