% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{solve_ip.min_cohorts_deps_installation_proposal}
\alias{solve_ip.min_cohorts_deps_installation_proposal}
\title{For each direct dependency, resolve that package using PPM snapshot as of release date + 1.
Finally, combine resolutions and run solve.}
\usage{
\method{solve_ip}{min_cohorts_deps_installation_proposal}(ip)
}
\description{
For each direct dependency, resolve that package using PPM snapshot as of release date + 1.
Finally, combine resolutions and run solve.
}
\keyword{internal}
