% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_ip.min_isolated_deps_installation_proposal}
\alias{solve_ip.min_isolated_deps_installation_proposal}
\title{Try to solve using min_isolated method. If Error - use \link{resolve_ignoring_release_remote}}
\usage{
\method{solve_ip}{min_isolated_deps_installation_proposal}(ip)
}
\description{
For each direct dependency, resolve that package using PPM snapshot as of release date + 1.
Finally, combine resolutions and run solve.
}
\keyword{internal}
