% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_individual_isolated.R
\name{test_install_deps_individually}
\alias{test_install_deps_individually}
\title{Test the install of all possible versions of the dependencies in isolation.}
\usage{
test_install_deps_individually(path)
}
\arguments{
\item{path}{(\code{character}) File path to a valid package directory that
contains a DESCRIPTION file.}
}
\description{
It looks at all dependencies individually and tries to install all possible
versions that exist on CRAN. This is a support function to help assess the
best minimal version of a package that builds in the environment.
}
\details{
Please note that because a given version of a package can be installed, it
does not mean that the API is compatible nor that it will be this version
that is selected.

Please refer to the documentation on the existing strategies
for more information (e.g. \link{new_max_deps_installation_proposal})

There is no output for this function other than the messages in the console.
}
\examples{
\dontshow{if (Sys.getenv("R_USER_CACHE_DIR", "") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
desc_path <- verdepcheck:::local_description(
  list(`dplyr (>= 1.1.2)` = "Import")
)
verdepcheck:::test_install_deps_individually(desc_path)
\dontshow{\}) # examplesIf}
}
