% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_individual_isolated.R
\name{test_install_pkg_versions}
\alias{test_install_pkg_versions}
\title{Test install all the versions of the package locally}
\usage{
test_install_pkg_versions(
  package,
  op = "",
  op_ver = "",
  max_count = NA,
  until_success = FALSE
)
}
\arguments{
\item{package}{(\code{character}) name of R package.}

\item{op}{(\code{character}) relational operator.}

\item{op_ver}{(\code{character}) version of package.}

\item{max_count}{(\code{integer}) maximum number of versions to test, \code{NA} for no
limit.}

\item{until_success}{(\code{logical}) flag that indicates whether a
successful installation of a package should stop the search.}
}
\description{
Support function to assess if every version specified by \code{op} and \code{op_ver}
of the package are installable. This is not always the case as there might
be some API changes that break specific versions.
}
\details{
The \code{Rcpp} package before 0.12.16 cannot be compiled by R versions above 4.0

There is no output for this function other than the messages in the console.
}
\examples{
test_install_pkg_versions("checkmate", ">=", "2.1.0")
test_install_pkg_versions("Rcpp", ">=", "0.12.15", 2)
test_install_pkg_versions("Rcpp", ">=", "0.12.15", NA, TRUE)
}
