% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{version_from_desc}
\alias{version_from_desc}
\title{Get package version from description}
\usage{
version_from_desc(pkg_name, desc)
}
\description{
Get package version from description
}
\examples{
d <- desc::desc(cmd = "!new")

d$set_dep("magrittr", type = "Imports", version = "*")
verdepcheck:::version_from_desc("magrittr", d)

d$set_dep("magrittr", type = "Imports", version = ">= 1.5")
verdepcheck:::version_from_desc("magrittr", d)
}
\keyword{internal}
