% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min_ver.R
\name{check_if_on_cran}
\alias{check_if_on_cran}
\title{Check if package is available on CRAN.}
\usage{
check_if_on_cran(remote_ref)
}
\description{
Check if package is available on CRAN.
}
\keyword{internal}
