% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{desc_add_extra_deps}
\alias{desc_add_extra_deps}
\title{Adds extra dependencies to the \code{desc} object.}
\usage{
desc_add_extra_deps(d, x)
}
\arguments{
\item{d}{(\code{desc}) DESCRIPTION object from \link[desc:desc]{desc::desc}}

\item{x}{(\code{character(1)}) Extra dependencies specified similarly as in the \code{DESCRIPTION} file, i.e.
\code{"<package name> (<operator> <version>)"} where both \verb{<operator>} and \verb{<version>} are optional.
Multiple entries are possible separated by \code{";"}.}
}
\value{
\code{desc} object with added dependencies.
}
\description{
This will add dependencies to the \code{Imports} field.
}
\examples{
d <- desc::desc(cmd = "!new")
d <- verdepcheck:::desc_add_extra_deps(d, "foo (>= 1.2.3)")
d <- verdepcheck:::desc_add_extra_deps(d, "bar (>= 2.3.4); baz (>= 3.4.5)")
d$get_deps()
}
\keyword{internal}
