% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{desc_remotes_cleanup}
\alias{desc_remotes_cleanup}
\title{Replace Remotes in the \code{desc} that have been resolved to a GitHub tag or are
in CRAN}
\usage{
desc_remotes_cleanup(d, new_refs)
}
\arguments{
\item{d}{(\code{desc}) DESCRIPTION object}

\item{new_refs}{(\code{list}) remote references that have been resolved and are
being updated in \code{Config/Needs/verdepcheck}}
}
\description{
Replaces any existing Remotes entry with the resolved GitHub tag from the
\code{new_refs}.
It keeps all the existing Remotes that have not been resolved in \code{new_refs}.
}
\keyword{internal}
