% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_desc_from_gh}
\alias{get_desc_from_gh}
\title{Get DESCRIPTION from GitHub Repository}
\usage{
get_desc_from_gh(org, repo, ref = "")
}
\description{
Get DESCRIPTION from GitHub Repository
}
\examples{
verdepcheck:::get_desc_from_gh("tidyverse", "tibble")
verdepcheck:::get_desc_from_gh("insightsengineering", "formatters", "v0.5.0")
}
\keyword{internal}
