% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_ref_max}
\alias{get_ref_max}
\title{Get reference to the maximal version of the package.}
\usage{
get_ref_max(remote_ref)
}
\arguments{
\item{remote_ref}{(\code{remote_ref}) object created with \code{\link[pkgdepends:parse_pkg_refs]{pkgdepends::parse_pkg_ref()}}}
}
\value{
(\code{remote_ref}) object with the package reference
}
\description{
Get reference to the maximal version of the package.
}
