% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_release_date}
\alias{get_release_date}
\title{Get release date.}
\usage{
get_release_date(remote_ref)
}
\arguments{
\item{remote_ref}{(\code{remote_ref}) object created with \code{\link[pkgdepends:parse_pkg_refs]{pkgdepends::parse_pkg_ref()}}}
}
\value{
(\code{remote_ref}) object with the package reference
}
\description{
Get release date.
}
