% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{version_from_desc}
\alias{version_from_desc}
\title{Get package version from description}
\usage{
version_from_desc(d, pkg_name)
}
\arguments{
\item{d}{(\code{desc}) DESCRIPTION object from \link[desc:desc]{desc::desc}}

\item{pkg_name}{(\code{character}) Package name}
}
\description{
Get package version from description
}
\examples{
d <- desc::desc(cmd = "!new")

d$set_dep("magrittr", type = "Imports", version = "*")
verdepcheck:::version_from_desc(d, "magrittr")

d$set_dep("magrittr", type = "Imports", version = ">= 1.5")
verdepcheck:::version_from_desc(d, "magrittr")
}
\keyword{internal}
